fix(ZMSKVR-405): allow editing future appointments while preventing calls
- Move date validation from WorkstationProcessGet to WorkstationProcess
- Future appointments can now be edited but cannot be called
- Improve exception templates with better navigation buttons:
  - Add 'Bearbeiten' button to redirect with selectedprocess parameter
  - Add 'Wechseln' button to workstation selection for scope errors
- Maintain all existing security validations for status and scope access
- Users can now edit future appointments normally in the interface
