+++ zmsadmin/templates/exception/bo/zmsapi/exception/process/processnotcurrentdate.twig
@@ -3,15 +3,18 @@
     {% block board_title %}Es ist ein Fehler aufgetreten.{% endblock %}
     {% block board_body %}
         <h2 class="exceptionData-headline">
-            Termin nicht verfügbar{% if sourceData.processId %} (Vorgang {{ sourceData.processId }}){% endif %}
+            Termin nicht aufrufbar{% if sourceData.processId %} (Vorgang {{ sourceData.processId }}){% endif %}
         </h2>
         <p>
-            {{ exclamationIcon|raw }} {{ errorPrefix }}: {% trans %}Der angegebene Termin liegt nicht am heutigen Tag und kann nicht aufgerufen werden. Sie können nur Termine des aktuellen Tages aufrufen.  Bitte prüfen Sie die angegebene Vorgangsnummer in der URL.{% endtrans %}
+            {{ exclamationIcon|raw }} {{ errorPrefix }}: {% trans %}Der angegebene Termin liegt nicht am heutigen Tag und kann nicht aufgerufen werden. Sie können nur Termine des aktuellen Tages aufrufen. Bitte prüfen Sie die angegebene Vorgangsnummer in der URL.{% endtrans %}
             {% if sourceData.appointmentDate %}
                 <br><br><strong>Terminzeit:</strong> {{ sourceData.appointmentDate }}{% if sourceData.appointmentTime %} um {{ sourceData.appointmentTime }}{% endif %}
             {% endif %}
         </p>
         <div class="form-actions">
+            {% if sourceData.processId %}
+                <button class="btn button-primary right" onclick="window.location.href='{{ urlGet('workstation') }}?selectedprocess={{ sourceData.processId }}'">Bearbeiten</button>
+            {% endif %}
             <button class="btn button-cancel right" onclick="window.location.href='{{ urlGet('workstation') }}'">Abbrechen</button>
         </div>
     {% endblock %}
